alert-aggregator/requirements.txt
kafka-python==2.0.2
elasticsearch==7.17.9
